Develop

+++ src/main/java/com/kks/demo/repository/InfoRepository.java
@@ -5,6 +5,7 @@ import com.kks.demo.dto.Follow;
 import com.kks.demo.dto.MyRecord;
 import com.kks.demo.dto.login.UserResponseDto;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -104,6 +105,20 @@ public class InfoRepository {
         return 1;
     }
 
+    /**
+     * 팔로우 여부
+     * */
+    public int getFollowStatus(String userId, String followId){
+        try{
+            return this.jdbcTemplate.queryForObject("select count(*) "
+                            + "from Follow "
+                            + "where followerIdx = " + userId + " AND followingIdx = " + followId,
+                    Integer.class);
+        }catch (EmptyResultDataAccessException e){
+            return 0;
+        }
+    }
+
 
     /**
      * 좋아요 게시물 목록
@@ -113,7 +128,7 @@ public class InfoRepository {
                         + "from   RecordLike rl "
                         + "inner join Record r "
                         + "		on rl.recordIdx = r.recordIdx "
-                        + "where  rl.userId = " + userId,
+                        + "where  rl.userId = " + userId + " and status = 'ACTIVE'",
                 (rs, rowNum) -> new MyRecord(
                         rs.getInt("recordIdx"),
                         rs.getString("imgUrl"))
@@ -125,43 +140,29 @@ public class InfoRepository {
      * */
     //사용자 계정 삭제
     public int deleteUser(String userId){
-        String deleteQuery = "delete from User where userId = ?";
-        Object[] deleteParams = new Object[]{userId};
-        this.jdbcTemplate.update(deleteQuery, deleteParams);
-
-        return 1;
+        String deleteQuery = "delete from User where userId = '" + userId + "'";
+        return this.jdbcTemplate.update(deleteQuery);
     }
 
     //팔로우 삭제
     public int deleteFollow(String userId){
-        String deleteQuery = "delete from Follow where followerIdx = ? and followingIdx = ?";
-        Object[] deleteParams = new Object[]{userId, userId};
-        this.jdbcTemplate.update(deleteQuery,deleteParams);
-
-        return 1;
+        String deleteQuery = "delete from Follow where followerIdx = '" + userId
+                + "' and followingIdx = '" + userId + "'";
+        return this.jdbcTemplate.update(deleteQuery);
     }
     //게시글 삭제
     public int deleteRecord(String userId){
-        String deleteQuery = "delete from Record where userId = ?";
-        Object[] deleteParams = new Object[]{userId, userId};
-        this.jdbcTemplate.update(deleteQuery,deleteParams);
-
-        return 1;
+        String deleteQuery = "delete from Record where userId = '" + userId + "'";
+        return this.jdbcTemplate.update(deleteQuery);
     }
     //좋아요 삭제
     public int deleteLike(String userId){
-        String deleteQuery = "delete from RecordLike where userId = ?";
-        Object[] deleteParams = new Object[]{userId, userId};
-        this.jdbcTemplate.update(deleteQuery,deleteParams);
-
-        return 1;
+        String deleteQuery = "delete from RecordLike where userId = '" + userId + "'";
+        return this.jdbcTemplate.update(deleteQuery);
     }
     //댓글 삭제
     public int deleteComment(String userId){
-        String deleteQuery = "delete from Comment where userId = ?";
-        Object[] deleteParams = new Object[]{userId, userId};
-        this.jdbcTemplate.update(deleteQuery,deleteParams);
-
-        return 1;
+        String deleteQuery = "delete from Comment where userId = '" + userId + "'";
+        return this.jdbcTemplate.update(deleteQuery);
     }
 }
